tango_with_django_project/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-bootstrap-toolkit==2.15.0
django-registration-redux==1.4
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0